+++ server/utils/s3.go
@@ -54,6 +54,18 @@ func createStorageClient() interface{} {
 	}
 }
 
+func UploadLogo(appName string, file *multipart.FileHeader, c *gin.Context, env *viper.Viper) (string, error) {
+	logoLink, _, err := UploadToS3(map[string]interface{}{
+		"app_name": appName,
+		"version":  "0.0.0",
+		"type":     "logo",
+		"channel":  "",
+		"platform": "",
+		"arch":     "",
+	}, file, c, env)
+	return logoLink, err
+}
+
 func UploadToS3(ctxQuery map[string]interface{}, file *multipart.FileHeader, c *gin.Context, env *viper.Viper) (string, string, error) {
 	// // Create an S3 client using another func
 	storageClient := createStorageClient()